refactor: ui_encoders > ook_tx, proc_ook > proc_ooktx
