+++ firmware/application/ui_navigation.cpp
@@ -41,7 +41,7 @@
 #include "ui_bht_tx.hpp"
 #include "ui_coasterp.hpp"
 #include "ui_debug.hpp"
-#include "ui_encoders.hpp"
+#include "ui_ook_tx.hpp"
 #include "ui_fileman.hpp"
 #include "ui_freqman.hpp"
 #include "ui_jammer.hpp"
